add additional tile types

+++ storage/entities.go
@@ -7,7 +7,7 @@ import "github.com/kpfaulkner/quadmap/quadmap"
 // QuadKey is NOT the primary key... but will be indexed and will be main column we search on.
 type TileEntity struct {
 	QuadKey     quadmap.QuadKey `db:"quadkey"`
-	TileType    uint16          `db:"tiletype"`
+	TileType    uint64          `db:"tiletype"`
 	Full        bool            `db:"full"`
 	DetailsMask uint64          `db:"details_mask"`
 	DetailsID   int64           `db:"details_id"`
@@ -18,7 +18,7 @@ type DetailsEntity struct {
 	Border          string `db:"border"`
 	SimpleBorder    string `db:"simple_border"`
 	SimpleBorderWKB []byte `db:"simple_border_wkb"`
-	TileType        uint16 `db:"tiletype"`
+	TileType        uint64 `db:"tiletype"`
 	DateTime        int64  `db:"datetime"`
 	Enabled         bool   `db:"enabled"`
 	Identifier      string `db:"identifier"`
